# Pre-emptively pause long resumed partitions to prevent read starvation