--- a/zio-kafka-test/src/test/scala/zio/kafka/consumer/internal/PollHistorySpec.scala
+++ b/zio-kafka-test/src/test/scala/zio/kafka/consumer/internal/PollHistorySpec.scala
@@ -1,7 +1,6 @@
 package zio.kafka.consumer.internal
 
 import zio.Scope
-import zio.kafka.consumer.internal.PollHistory.PollHistoryImpl
 import zio.test._
 
 object PollHistorySpec extends ZIOSpecDefault {
@@ -33,16 +32,4 @@ object PollHistorySpec extends ZIOSpecDefault {
       )
     }
   )
-
-  private implicit class RichPollHistory(private val ph: PollHistory) extends AnyVal {
-    def asBitString: String =
-      ph.asInstanceOf[PollHistoryImpl].resumeBits.toBinaryString
-  }
-
-  private implicit class PollHistoryOps(private val s: String) extends AnyVal {
-    def toPollHistory: PollHistory =
-      s.foldLeft(PollHistory.Empty) { case (ph, b) =>
-        ph.addPollHistory(b == '1')
-      }
-  }
 }