--- a/zio-kafka/src/main/scala-2.12/zio/kafka/consumer/internal/internal.scala
+++ b/zio-kafka/src/main/scala-2.12/zio/kafka/consumer/internal/internal.scala
@@ -2,10 +2,10 @@ package zio.kafka.consumer
 
 import scala.collection.SeqView
 
+/** Provides scala 2.13 forward compatibility for scala 2.12. */
 package object internal {
 
-  /** Provide scala 2.13 forward compatibility for the 2.12 build. */
-  implicit class ExtraIterableOnceOpsOps[A, Coll](val s: SeqView[A, Coll]) {
+  implicit class ExtraSeqViewOps[A, Coll](val s: SeqView[A, Coll]) {
     def maxByOption[B](f: A => B)(implicit cmp: Ordering[B]): Option[A] =
       if (s.isEmpty) None
       else Some(s.maxBy(f)(cmp))