The Kafka client has a tendency to consume everything from 1 partition before going to the next one. This can cause a partition from not being read for a long time, or even forever. To prevent this, when multiple partitions are resumed, we force pause the longest resumed partition (though we allow it to be resumed for 4 polls).

Zio-kafka version 2.3 introduced optimistic pre-fetching. This causes partitions to be unpaused much more often. Because of this, mentioned Kafka behavior became more apparent for high load applications.

Todo before this PR goes out of draft:
 - [x] make algorithm less eager
 - [x] write tests
 - [ ] Fix performance regression (see https://github.com/zio/zio-kafka/pull/844#issuecomment-1573545507)
 - [ ] do not use package object (will be dropped in scala 3)

Also, since this might be related, we need to see if we can fix the increase of CPU usage and GC pressure of zio-kafka 2.3.