@guizmaii 2023-05-12
I'm not sure we want to manipulate this behaviour. If Kafka is doing that, it's probably because it optimizes the consumption, the interaction between the client and the server and the loading of data from the disk on the server side