@erikvanoosten 2023-05-12
The Kafka client optimizes for throughput. Some people need low latency, or a balance between latency and throughput.