@svroonland 2023-05-13
Could we write a test case demonstrating this behavior?