@erikvanoosten 2023-05-13
> Could we write a test case demonstrating this behavior?

By 'this behavior' do you mean the tendency of the client to consume one partition as long as possible, or do you mean a test that validates that the proposed implementation works as expected (or something else)?