@erikvanoosten 2023-05-13
Earlier I wrote:
> the akka streaming client does something like this.

This statement was based on anecdotal evidence 😉 , however, now I look in the Alpakka code, I cannot find such a feature. So now I am not sure where I heard this.