@erikvanoosten 2023-05-13
> The former

We have observed this at work. You can see this by tracking lag of the different partitions.

We're not the only the see this. A user reported a few days ago that this is happening to them on zio-kafka: https://discord.com/channels/629491597070827530/629497941719121960/1105883502710366228

Also this is interesting: https://stackoverflow.com/questions/76175977/how-to-make-consumer-to-switch-between-partitions-in-round-robbin-fashion