@erikvanoosten 2023-05-14
> The implementation in this PR is very eager. We can easily tune it down a bit. For example, we could delay force-pausing a partition after at least n resume polls. I doubt anyone will suffer from this.

This has now been implemented.