@guizmaii 2023-06-02
Benchmarks results:

Before
```scala
zioKafkaBench/Jmh/run -wi 10 -i 10 -r 1 -w 1 -t 1 -f 3 -foe true
...
[info] Result "zio.kafka.bench.ConsumersComparisonBenchmark.zioKafka":
[info]   739.566 ±(99.9%) 21.019 ms/op [Average]
[info]   (min, avg, max) = (682.067, 739.566, 813.396), stdev = 31.461
[info]   CI (99.9%): [718.547, 760.586] (assumes normal distribution)
[info] # Run complete. Total time: 00:18:24
...
[info] Benchmark                                        Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput                     avgt   30   205.133 ±   6.544  ms/op
[info] ConsumerBenchmark.throughputWithCommits          avgt   30   284.912 ±   5.204  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients        avgt   30   645.361 ±   3.004  ms/op
[info] ConsumersComparisonBenchmark.manualKafkaClients  avgt   30   641.042 ±   2.877  ms/op
[info] ConsumersComparisonBenchmark.manualZioKafka      avgt   30  2651.134 ± 100.297  ms/op
[info] ConsumersComparisonBenchmark.zioKafka            avgt   30   739.566 ±  21.019  ms/op
```

After
```scala
zioKafkaBench/Jmh/run -wi 10 -i 10 -r 1 -w 1 -t 1 -f 3 -foe true
...
[info] Result "zio.kafka.bench.ConsumersComparisonBenchmark.zioKafka":
[info]   1499.816 ±(99.9%) 216.967 ms/op [Average]
[info]   (min, avg, max) = (1002.257, 1499.816, 2254.385), stdev = 324.746
[info]   CI (99.9%): [1282.849, 1716.783] (assumes normal distribution)
[info] # Run complete. Total time: 00:18:15
...
[info] Benchmark                                        Mode  Cnt     Score     Error  Units
[info] ConsumerBenchmark.throughput                     avgt   30   222.238 ±  13.788  ms/op
[info] ConsumerBenchmark.throughputWithCommits          avgt   30   283.729 ±   6.562  ms/op
[info] ConsumersComparisonBenchmark.kafkaClients        avgt   30   645.774 ±   3.357  ms/op
[info] ConsumersComparisonBenchmark.manualKafkaClients  avgt   30   648.834 ±   3.384  ms/op
[info] ConsumersComparisonBenchmark.manualZioKafka      avgt   30  2412.388 ± 350.380  ms/op
[info] ConsumersComparisonBenchmark.zioKafka            avgt   30  1499.816 ± 216.967  ms/op
```

This might be expected, but there's a clear performance regression.
I think these results are another argument for me to ask for this behaviour to be put behind an option, which would be not activated by default.

if this is an optimisation for latency, I'd see something like:
```scala
sealed trait Mode
object Mode {
  case object Throughput extends Mode
  case object Latency    extends Mode
}

object Consumer {
  def make(
    ...
    mode: Mode = Mode.Throughput // To keep the previous behaviour as the default one
  )
}
```

WDYT?