@guizmaii 2023-06-02
@erikvanoosten 2 questions about your comment https://github.com/zio/zio-kafka/pull/844#issuecomment-1570223553

1. You say:
> CPU and GC pressure is slightly elevated but this seems to be a ZIO issue (see https://github.com/zio/zio/issues/8158)

Now that ZIO 2.0.14 containing the mentioned fix is released. Can you confirm that this was indeed because of the issue in ZIO, please?

2. When you say:
> All other metrics like heap usage, latency, consumer lag, etc. are the same as before.

What do you mean by "before"? Which version of zio-kafka are you talking about 2.3.x or 2.2? People are reporting high CPU and memory usage since 2.3.0.