@guizmaii 2023-06-02
I think before going forward with these changes, I'd like to introduce the option to disable the optimistic resume so that users like @myazinn who are experiencing issues (high CPU and memory usage, high lag) with 2.3.x can disable it and tell us if it fixes their issue.

WDYT?
Cc @svroonland