@erikvanoosten 2023-06-02
> What do you mean by "before"? Which version of zio-kafka are you talking about 2.3.x or 2.2? People are reporting high CPU and memory usage since 2.3.0.

The version I am running is a fork of 2.2. So this is comparing 2.2 with and without the feature this PR introduces.