@erikvanoosten 2023-06-02
> > CPU and GC pressure is slightly elevated but this seems to be a ZIO issue (see https://github.com/zio/zio/issues/8158)

> Now that ZIO 2.0.14 containing the mentioned fix is released. Can you confirm that this was indeed because of the issue in ZIO, please?

Sorry, no I cannot confirm that. Worse, I have to (mostly) withdraw my previous observation as well.

I took a second look at the changes from introducing this feature and also at the changes from updating to zio 2.0.14. Now that I am properly averaging the CPU and GC metrics over all the pods, and also, look at them over a longer period (days instead of a few hours) I see that these metrics are too volatile to attribute differences to these features.

The only metric that did change is latency (time from produce in source to end of processing). That metric was not affected by this PR. Upgrading to zio 2.0.14 however had a very strong positive effect.

Please note that these observations are all based on zio-kafka 2.2.