@svroonland 2023-06-03
I can attest that partition skew is not nice to have in streaming applications, where there's usually some latency constraints. Was not aware that kafka itself can cause this, not just a skewed number of messages on partitions. I believe at the core this is a kafka issue that should be resolved in the kafka client or broker. So about the KIP-41 that mentions this problem, was that never implemented? Can we file a new issue somewhere?

As an alternative for this PR:

If I remember correctly, in a project I did a few years ago we had a large variation in message processing times due to the varying nature of the messages. We kept track of the latest message timestamp on the partitions as well as the timestamp of the last processed messages. If the diff for one of the partitions got too large, we would pause the other ones. 

We could implement something similar as a separate component that you can pass your partition streams through, eg a function `Stream[Throwable, (TopicPartition, ZStream[R, Throwable, CommittableRecord[K, V]])] = Stream[Throwable, (TopicPartition, ZStream[R, Throwable, CommittableRecord[K, V]])]`.